Move evaluators into mir_eval
