+++ evaluators/melody_eval.py
@@ -15,69 +15,7 @@ J. Salamon, E. Gomez, D. P. W. Ellis and G. Richard, "Melody Extraction
 from Polyphonic Music Signals: Approaches, Applications and Challenges",
 IEEE Signal Processing Magazine, 31(2):118-134, Mar. 2014.
 '''
-
-from __future__ import print_function
-import argparse
-import sys
-import os
-
 import mir_eval
 
-from . import eval_utilities
-
-
-def process_arguments():
-    '''Argparse function to get the program parameters'''
-
-    parser = argparse.ArgumentParser(description='mir_eval melody extraction '
-                                                 'evaluation')
-
-    parser.add_argument('-o',
-                        dest='output_file',
-                        default=None,
-                        type=str,
-                        action='store',
-                        help='Store results in json format')
-
-    parser.add_argument('reference_file',
-                        action='store',
-                        help='path to the ground truth annotation')
-
-    parser.add_argument('estimated_file',
-                        action='store',
-                        help='path to the estimation file')
-
-    parser.add_argument("--hop",
-                        dest='hop',
-                        type=float,
-                        default=None,
-                        help="hop size (in seconds) to use for the evaluation"
-                        " (optional)")
-
-    return vars(parser.parse_args(sys.argv[1:]))
-
-
-def main():
-    # Get the parameters
-    parameters = process_arguments()
-
-    # Load in the data from the provided files
-    (ref_time,
-     ref_freq) = mir_eval.io.load_time_series(parameters['reference_file'])
-    (est_time,
-     est_freq) = mir_eval.io.load_time_series(parameters['estimated_file'])
-
-    # Compute all the scores
-    scores = mir_eval.melody.evaluate(ref_time, ref_freq, est_time, est_freq,
-                                      hop=parameters['hop'])
-    print("{} vs. {}".format(os.path.basename(parameters['reference_file']),
-                             os.path.basename(parameters['estimated_file'])))
-    eval_utilities.print_evaluation(scores)
-
-    if parameters['output_file']:
-        print('Saving results to: ', parameters['output_file'])
-        eval_utilities.save_results(scores, parameters['output_file'])
-
-
 if __name__ == '__main__':
-    main()
+    mir_eval.melody.main()
