+++ evaluators/segment_hier_eval.py
@@ -11,89 +11,7 @@ Usage:
                        [-o output.json] \
                        [-w WINDOW_SIZE]
 '''
-
-from __future__ import print_function
-import argparse
-import sys
-from os.path import basename
-
 import mir_eval
-from mir_eval.io import load_labeled_intervals
-
-from . import eval_utilities
-
-
-def process_arguments():
-    '''Argparse function to get the program parameters'''
-
-    parser = argparse.ArgumentParser(description='mir_eval hierarchical '
-                                     'segmentation evaluation')
-
-    parser.add_argument('-w',
-                        '--window',
-                        dest='window',
-                        default="15.0",
-                        type=float,
-                        help='Window length for t-measures')
-
-    parser.add_argument('-o',
-                        dest='output_file',
-                        default=None,
-                        type=str,
-                        action='store',
-                        help='Store results in json format')
-
-    parser.add_argument('-r',
-                        '--reference',
-                        dest='reference_file',
-                        nargs='+',
-                        type=str,
-                        action='store',
-                        help='path to the reference annotation(s) in '
-                        '.lab format, ordered from top to bottom of '
-                        'the hierarchy')
-
-    parser.add_argument('-e',
-                        '--estimate',
-                        dest='estimated_file',
-                        nargs='+',
-                        type=str,
-                        action='store',
-                        help='path to the estimated annotation(s) in '
-                        '.lab format, ordered from top to bottom of '
-                        'the hierarchy')
-
-    return vars(parser.parse_args(sys.argv[1:]))
-
-
-def main():
-    # Get the parameters
-    parameters = process_arguments()
-
-    # load the data
-    ref_files = parameters['reference_file']
-    est_files = parameters['estimated_file']
-
-    ref = [load_labeled_intervals(_) for _ in ref_files]
-    est = [load_labeled_intervals(_) for _ in est_files]
-    ref_intervals = [seg[0] for seg in ref]
-    ref_labels = [seg[1] for seg in ref]
-    est_intervals = [seg[0] for seg in est]
-    est_labels = [seg[1] for seg in est]
-
-    # Compute all the scores
-    scores = mir_eval.hierarchy.evaluate(ref_intervals, ref_labels,
-                                         est_intervals, est_labels,
-                                         window=parameters['window'])
-    print("{} [...] vs. {} [...]".format(
-        basename(parameters['reference_file'][0]),
-        basename(parameters['estimated_file'][0])))
-    eval_utilities.print_evaluation(scores)
-
-    if parameters['output_file']:
-        print('Saving results to: ', parameters['output_file'])
-        eval_utilities.save_results(scores, parameters['output_file'])
-
 
 if __name__ == '__main__':
-    main()
+    mir_eval.hierarchy.main()
